Backend/neuronflow-agent/pyproject.toml
[project]
name = "neuronflow-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "livekit-agents>=1.4.4",
    "livekit-plugins-aws[realtime]>=1.4.4",
    "livekit-plugins-silero>=1.4.4",
    "python-dotenv>=1.2.2",
]
